+++ server/config/config.go
@@ -21,17 +21,13 @@ import (
 	"path"
 	"time"
 
-	sdkmath "cosmossdk.io/math"
-
 	"github.com/spf13/viper"
 
 	"github.com/cometbft/cometbft/libs/strings"
 
 	errorsmod "cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/server/config"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/rosetta"
 )
 
 const (
@@ -52,8 +48,6 @@ const (
 
 	// DefaultFixRevertGasRefundHeight is the default height at which to overwrite gas refund
 	DefaultFixRevertGasRefundHeight = 0
-
-	DefaultMaxTxGasWanted = 0
 
 	DefaultGasCap uint64 = 25000000
 
@@ -83,30 +77,12 @@ const (
 	// DefaultReturnDataLimit is maximum number of bytes returned from eth_call or similar invocations
 	DefaultReturnDataLimit = 100000
 
-	// DefaultRosettaEnable is the default value for the parameter that defines if the Rosetta API server is enabled
-	DefaultRosettaEnable = false
-
-	// DefaultRosettaBlockchain defines the default blockchain name for the rosetta server
-	DefaultRosettaBlockchain = "evmos"
-
-	// DefaultRosettaNetwork defines the default network name for the rosetta server
-	DefaultRosettaNetwork = "evmos"
-
-	// DefaultRosettaGasToSuggest defines the default gas to suggest for the rosetta server
-	DefaultRosettaGasToSuggest = 300_000
-
-	// DefaultRosettaDenomToSuggest defines the default denom for fee suggestion
-	DefaultRosettaDenomToSuggest = "basecro"
-
 	BlockExecutorSequential = "sequential"
 	BlockExecutorBlockSTM   = "block-stm"
 	DefaultMaxTxs           = 3000
 )
 
 var (
-	// DefaultRosettaGasPrices defines the default list of prices to suggest
-	DefaultRosettaGasPrices = sdk.NewDecCoins(sdk.NewDecCoin(DefaultRosettaDenomToSuggest, sdkmath.NewInt(4_000_000)))
-
 	evmTracers = []string{"json", "markdown", "struct", "access_list"}
 
 	blockExecutors = []string{BlockExecutorSequential, BlockExecutorBlockSTM}
@@ -120,7 +96,6 @@ type Config struct {
 	EVM     EVMConfig     `mapstructure:"evm"`
 	JSONRPC JSONRPCConfig `mapstructure:"json-rpc"`
 	TLS     TLSConfig     `mapstructure:"tls"`
-	Rosetta RosettaConfig `mapstructure:"rosetta"`
 }
 
 // EVMConfig defines the application configuration values for the EVM.
@@ -128,8 +103,6 @@ type EVMConfig struct {
 	// Tracer defines vm.Tracer type that the EVM will use if the node is run in
 	// trace mode. Default: 'json'.
 	Tracer string `mapstructure:"tracer"`
-	// MaxTxGasWanted defines the gas wanted for each eth tx returned in ante handler in check tx mode.
-	MaxTxGasWanted uint64 `mapstructure:"max-tx-gas-wanted"`
 	// BlockExecutor set block executor type, "block-stm" for parallel execution, "sequential" for sequential execution.
 	BlockExecutor string `mapstructure:"block-executor"`
 	// BlockSTMWorkers is the number of workers for block-stm execution, `0` means using all available CPUs.
@@ -195,13 +168,6 @@ type TLSConfig struct {
 	KeyPath string `mapstructure:"key-path"`
 }
 
-// RosettaConfig defines configuration for the Rosetta server.
-type RosettaConfig struct {
-	rosetta.Config
-	// Enable defines if the Rosetta server should be enabled.
-	Enable bool `mapstructure:"enable"`
-}
-
 // AppConfig helps to override default appConfig template and configs.
 // return "", nil if no custom configuration is required for the application.
 func AppConfig(denom string) (string, interface{}) {
@@ -244,16 +210,14 @@ func DefaultConfig() *Config {
 		EVM:     *DefaultEVMConfig(),
 		JSONRPC: *DefaultJSONRPCConfig(),
 		TLS:     *DefaultTLSConfig(),
-		Rosetta: *DefaultRosettaConfig(),
 	}
 }
 
 // DefaultEVMConfig returns the default EVM configuration
 func DefaultEVMConfig() *EVMConfig {
 	return &EVMConfig{
-		Tracer:         DefaultEVMTracer,
-		MaxTxGasWanted: DefaultMaxTxGasWanted,
-		BlockExecutor:  BlockExecutorSequential,
+		Tracer:        DefaultEVMTracer,
+		BlockExecutor: BlockExecutorSequential,
 	}
 }
 
@@ -366,26 +330,6 @@ func DefaultTLSConfig() *TLSConfig {
 	}
 }
 
-// DefaultEVMConfig returns the default EVM configuration
-func DefaultRosettaConfig() *RosettaConfig {
-	return &RosettaConfig{
-		Config: rosetta.Config{
-			Blockchain:          DefaultRosettaBlockchain,
-			Network:             DefaultRosettaNetwork,
-			TendermintRPC:       rosetta.DefaultCometEndpoint,
-			GRPCEndpoint:        rosetta.DefaultGRPCEndpoint,
-			Addr:                rosetta.DefaultAddr,
-			Retries:             rosetta.DefaultRetries,
-			Offline:             rosetta.DefaultOffline,
-			EnableFeeSuggestion: rosetta.DefaultEnableFeeSuggestion,
-			GasToSuggest:        DefaultRosettaGasToSuggest,
-			DenomToSuggest:      DefaultRosettaDenomToSuggest,
-			GasPrices:           DefaultRosettaGasPrices,
-		},
-		Enable: DefaultRosettaEnable,
-	}
-}
-
 // Validate returns an error if the TLS certificate and key file extensions are invalid.
 func (c TLSConfig) Validate() error {
 	certExt := path.Ext(c.CertificatePath)
@@ -414,7 +358,6 @@ func GetConfig(v *viper.Viper) (Config, error) {
 		Config: cfg,
 		EVM: EVMConfig{
 			Tracer:              v.GetString("evm.tracer"),
-			MaxTxGasWanted:      v.GetUint64("evm.max-tx-gas-wanted"),
 			BlockExecutor:       v.GetString("evm.block-executor"),
 			BlockSTMWorkers:     v.GetInt("evm.block-stm-workers"),
 			BlockSTMPreEstimate: v.GetBool("evm.block-stm-pre-estimate"),