+++ server/config/toml.go
@@ -28,9 +28,6 @@ const DefaultConfigTemplate = `
 # Valid types are: json|struct|access_list|markdown
 tracer = "{{ .EVM.Tracer }}"
 
-# MaxTxGasWanted defines the gas wanted for each eth tx returned in ante handler in check tx mode.
-max-tx-gas-wanted = {{ .EVM.MaxTxGasWanted }}
-
 # BlockExecutor set block executor type, "block-stm" for parallel execution, "sequential" for sequential execution.
 block-executor = "{{ .EVM.BlockExecutor }}"
 # BlockSTMWorkers is the number of workers for block-stm execution, 0 means using all available CPUs.