+++ types/block.go
@@ -15,7 +15,11 @@
 // along with the Ethermint library. If not, see https://github.com/evmos/ethermint/blob/main/LICENSE
 package types
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	math "math"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
 
 // BlockGasLimit returns the max gas (limit) defined in the block gas meter. If the meter is not
 // set, it returns the max gas from the application consensus params.
@@ -28,6 +32,13 @@ func BlockGasLimit(ctx sdk.Context) uint64 {
 	}
 
 	maxGas := cp.Block.MaxGas
+
+	// Setting max_gas to -1 in CometBFT means there is no limit on the maximum gas consumption for transactions
+	// https://github.com/cometbft/cometbft/blob/v0.37.2/proto/tendermint/types/params.proto#L25-L27
+	if maxGas == -1 {
+		return math.MaxUint64
+	}
+
 	if maxGas > 0 {
 		return uint64(maxGas)
 	}